# src/common/client/openapi.json
{
  "openapi": "3.0.2",
  "info": { "title": "Druffle Development API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/room/services/{room_service_id}": {
      "get": {
        "tags": ["Room Services"],
        "summary": "Get Room Service By Id",
        "description": "Get room_service by id. Applicable for Administrative accounts",
        "operationId": "get_room_service_by_id_api_v1_room_services__room_service_id__get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Room Service Id", "type": "string" },
            "name": "room_service_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomServiceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Room Services"],
        "summary": "Update Room Service By Id",
        "description": "Update room_service by id. Applicable for Administrative accounts",
        "operationId": "update_room_service_by_id_api_v1_room_services__room_service_id__put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Room Service Id", "type": "string" },
            "name": "room_service_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RoomServiceUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomServiceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "delete": {
        "tags": ["Room Services"],
        "summary": "Delete Room Service By Id",
        "description": "Delete room_service by id. Applicable for Administrative accounts",
        "operationId": "delete_room_service_by_id_api_v1_room_services__room_service_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Room Service Id", "type": "string" },
            "name": "room_service_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomServiceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "RoomServiceResponse": {
        "title": "RoomServiceResponse",
        "required": ["is_active", "service_type", "hotel_id", "id"],
        "type": "object",
        "properties": {
          "is_deleted": { "title": "Is Deleted", "type": "boolean" },
          "description": { "title": "Description", "type": "string" },
          "custom_name": { "title": "Custom Name", "type": "string" },
          "price": { "title": "Price", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean" },
          "service_type": { "$ref": "#/components/schemas/RoomServicesTypes" },
          "hotel_id": { "title": "Hotel Id", "type": "string", "format": "uuid4" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "RoomServiceUpdateRequest": {
        "title": "RoomServiceUpdateRequest",
        "required": ["is_active"],
        "type": "object",
        "properties": {
          "description": { "title": "Description", "type": "string" },
          "custom_name": { "title": "Custom Name", "type": "string" },
          "price": { "title": "Price", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean" }
        }
      },
      "RoomServicesTypes": {
        "title": "RoomServicesTypes",
        "enum": ["order_food", "upgrade_room", "laundry", "spa", "other"],
        "description": "An enumeration."
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "JWTokenAuth": { "type": "apiKey", "in": "header", "name": "Authorization" }
    }
  }
}